Refactor priority search.
