+++ opto/trainer/algorithms/priority_search/__init__.py
@@ -1,0 +1,1 @@
+from opto.trainer.algorithms.priority_search.priority_search import PrioritySearch
