+++ opto/trainer/algorithms/priority_search/priority_search.py
@@ -1,306 +1,14 @@
 import numpy as np
 import copy
 import heapq
-from dataclasses import dataclass
+import time
 from typing import Union, List, Tuple, Dict, Any, Optional
 from opto import trace
 from opto.trace.nodes import ParameterNode
-from opto.trainer.utils import async_run, batch_run
-from opto.optimizers.utils import print_color
-from opto.trainer.algorithms.basic_algorithms import Minibatch, AlgorithmBase, batchify
-from opto.trainer.evaluators import evaluate
-from opto.trainer.loader import DataLoader
-from opto.trainer.sampler import Sampler, RolloutsGraph
-import time
-
-# TODO save and load SearchTemplate
-# TODO async version???
-# TODO create SYNC and ASYNC versions of the base class; add an attribute to the class to indicate
-# TODO a better data structure to store samples
-
-
-# Some helper functions to convert between trace.Module and update_dict
-
-def get_original_name(node):
-    """Extract the original name from a node, removing all _copy suffixes."""
-    py_name = node.py_name  # This removes colons: "param:0" -> "param0"
-
-    # Find the first occurrence of "_copy" and remove it and everything after
-    copy_index = py_name.find('_copy')
-    if copy_index != -1:
-        return py_name[:copy_index]
-    else:
-        return py_name
-
-def is_node_copy(a, b):
-    """Check if two nodes are copies of each other by comparing their original names.
-
-    This function has transitivity: if A is a copy of B and B is a copy of C,
-    then A is also considered a copy of C.
-    """
-    return get_original_name(a) == get_original_name(b)
-
-def is_module_copy(a, b):
-    """ Check if a and b (trace.Modules) are copies of each other. """
-    parameters_a = a.parameters() # list of ParameterNode
-    parameters_b = b.parameters() # list of ParameterNode
-    # Check if all parameters of a are copies of b or vice versa
-    # This might over count
-    # need to check 1:1 correspondence
-    matched = []
-    for p_a in parameters_a:
-        _matched = []
-        for p_b in parameters_b:
-            _matched.append(is_node_copy(p_a, p_b))
-    np.array(matched)
-    if np.all(np.sum(matched, axis=1) == 1) and np.all(np.sum(matched, axis=0) == 1):
-        return True
-    return False
-
-def remap_update_dict(base_module, update_dict):
-    """ Remap the update dict to the agent's parameters. update_dict might have keys which are copies of the base_module's parameters or visa versa.
-        This function remaps the keys in update_dict to the original parameters of the base_module.
-
-        The return dict is empty if no keys in update_dict matched any parameters of the base_module. This condition can be used to check if the update_dict contains non-trivial updates.
-    """
-    parameters = base_module.parameters()  # get the parameters of the base agent
-    remapped_update_dict = {}
-    for k, v in update_dict.items():
-        for p in parameters:
-            if is_node_copy(k, p): # Check if k is a copy of p or p is a copy of k
-                remapped_update_dict[p] = v
-                break # stop checking once we've found a match
-    return remapped_update_dict
-
-def set_module_parameters(agent, update_dict):
-    """ Set the parameters of the agent based on the update_dict.
-        The update_dict is a dictionary of ParameterNode: value pairs.
-        The agent's parameters will be updated with the values from the update_dict.
-    """
-    remapped_update_dict = remap_update_dict(agent, update_dict)  # remap the update dict to the agent's parameters
-    for k, v in remapped_update_dict.items():
-        k._data = v  # set the parameter's data to the value in the update_dict
-
-def create_module_from_update_dict(agent, update_dict):
-    """ Create a new agent from the update_dict.
-        The update_dict is a dictionary of ParameterNode: value pairs.
-        A new agent will be created with the parameters set to the values from the update_dict.
-    """
-    new_agent = copy.deepcopy(agent) #.copy()  # create a copy of the agent
-    set_module_parameters(new_agent, update_dict)  # set the parameters of the new agent
-    return new_agent  # return the new agent
-
-
-
-class Samples:
-    """ A container for samples collected during the search algorithm. It contains a list of RolloutsGraph objects
-    and a dataset with inputs and infos which created the list of RolloutsGraph. """
-
-    samples: List[RolloutsGraph]
-    dataset: Dict[str, List[Any]]  # contains 'inputs' and 'infos' keys
-
-    def __init__(self, samples: List[RolloutsGraph], dataset: Dict[str, List[Any]]):
-        assert isinstance(samples, list), "samples must be a list of RolloutsGraph objects."
-        assert all(isinstance(s, RolloutsGraph) for s in samples), "All samples must be RolloutsGraph objects."
-        assert isinstance(dataset, dict), "dataset must be a dict."
-        assert 'inputs' in dataset and 'infos' in dataset, "dataset must contain 'inputs' and 'infos' keys."
-
-        self.samples = samples
-        self.dataset = dataset  # NOTE this cannot be extracted from the samples in general?
-
-    def add_samples(self, samples):
-        """ Add samples to the Samples object. """
-        assert isinstance(samples, Samples), "samples must be an instance of Samples."
-        samples = samples.samples  # extract the samples from the Samples object
-        assert isinstance(samples, list), "samples must be a list of RolloutsGraph objects."
-        assert all(isinstance(s, RolloutsGraph) for s in samples), "All samples must be RolloutsGraph objects."
-
-        # TODO assert xs and infos are in self.minibatch
-        # add a function to extract unique inputs and infos from the samples
-
-        self.samples.extend(samples)
-
-    def get_batch(self):
-        return self.dataset #['inputs'], self.minibatch['infos']
-
-    def __iter__(self):
-        """ Iterate over the samples. """
-        return iter(self.samples)
-
-    def __len__(self):
-        return sum(len(s) for s in self.samples)
-
-
-
-class SearchTemplate(Minibatch):
-    # This only uses __init__ and evaluate of Minibatch class.
-    """ This implements a generic template for search algorithm. """
-
-    def train(self,
-              guide, # guide to provide feedback
-              train_dataset,  # dataset of (x, info) pairs to train the agent
-              *,
-              # validation
-              validate_dataset = None, # same format as train_dataset; if None use the current batch.
-              validate_guide = None,  #  to provide scores for the validation set
-              # training loop
-              batch_size = 1,  # batch size for updating the agent
-              sub_batch_size = None,  # sub-batch size for broadcasting the agents
-              score_range = None,  # minimum score to update the agent
-              num_epochs = 1,  # number of training epochs
-              num_threads = None,  # maximum number of threads to use
-              verbose = False,  # whether to print the output of the agent
-              # evaluation
-              test_dataset = None, # dataset of (x, info) pairs to evaluate the agent; if None, use train_dataset
-              test_guide = None, # guide to provide scores for the test set; if None, use guide
-              eval_frequency: Union[int, None] = 1,  # frequency of evaluation
-              num_eval_samples: int = 1,  # number of samples to use to evaluate each input
-              # logging
-              log_frequency = None,  # frequency of logging
-              save_frequency: Union[int, None] = None,  # frequency of saving the agent
-              save_path: str = "checkpoints/agent.pkl",  # path to save the agent
-              **kwargs
-              ):
-
-        ## Setup
-        test_frequency = eval_frequency  # use eval_frequency as test_frequency  # NOTE legacy notation
-        log_frequency = log_frequency or test_frequency  # frequency of logging (default to test_frequency)
-        self.num_threads = num_threads or self.num_threads  # Use provided num_threads or fall back to self.num_threads
-        test_dataset = test_dataset or train_dataset  # default to train_dataset if test_dataset is not provided
-        test_guide = test_guide or guide
-        self.num_eval_samples = num_eval_samples  # number of samples to use to evaluate each input
-        self.score_range = score_range or (-np.inf, np.inf)
-
-        self.train_sampler = Sampler(
-            DataLoader(train_dataset, batch_size=batch_size),
-            guide,
-            num_threads=self.num_threads,
-            sub_batch_size=sub_batch_size,
-            score_range=self.score_range
-        )
-        self._validate_dataset = validate_dataset  # if None, the current batch will be used for validation
-        self.validate_sampler = Sampler(
-            DataLoader(validate_dataset if validate_dataset else {'inputs':[],'infos':[]}, batch_size=batch_size),
-            validate_guide or guide,
-            num_threads=self.num_threads,
-            sub_batch_size=None,  # no sub-batch size for validation
-            score_range=self.score_range
-        )
-
-        # Evaluate the agent before learning
-        # NOTE set test_frequency < 0 to skip first evaluation
-        if (test_frequency is not None) and test_frequency > 0:
-            info_test = self.test(test_dataset, test_guide)  # test self.agent
-            self.log(info_test)
-
-        # Save the agent before learning if save_frequency > 0
-        if (save_frequency is not None) and save_frequency > 0:
-            self.save(save_path)
-
-        samples = None
-        self.n_epochs = 0 # number of epochs (full passes over the dataset) performed by the algorithm (This is incremented in sample)
-        self.n_samples = 0 # number of training samples processed by the algorithm (This is incremented in sample)
-        train_scores = []  # to store the scores of the agent during training
-
-        while self.n_epochs < num_epochs :
-
-            print(f"Epoch: {self.n_epochs}. Iteration: {self.n_iters}")
-
-            # 1. Propose new parameters given the current state of the algorithm
-            # proposals: list of trace.Modules
-            update_dict, proposals, info_update = self.update(samples, verbose=verbose, **kwargs)
-            self.optimizer.update(update_dict)  # update self.agent with the proposed parameters
-
-            # 2. Get feedback on the proposed parameters on the current batch
-            # samples: Samples object containing the samples and the minibatch
-            samples, info_sample = self.sample(proposals, verbose=verbose, **kwargs)
-
-            # Evaluate the agent after update
-            if (test_frequency is not None) and (self.n_iters % test_frequency == 0):
-                info_test = self.test(test_dataset, test_guide)  # test self.agent
-                self.log(info_test, prefix="Test: ")
-
-            # Save the algorithm state
-            if (save_frequency is not None and save_frequency > 0) and self.n_iters % save_frequency == 0:
-                self.save(save_path)
-
-            # Log information
-            assert 'mean_score' in info_sample, "info_sample must contain 'mean_score'."
-            assert 'n_epochs' in info_sample, "info_sample must contain 'n_epochs'."
-
-            train_scores.append(info_sample['mean_score'])  # so that mean can be computed
-            if self.n_iters % log_frequency == 0:
-                self.logger.log('Average train score', np.mean(train_scores), self.n_iters, color='blue')
-                self.log(info_update, prefix="Update: ")
-                self.log(info_sample, prefix="Sample: ")
-                self.n_samples += len(samples)  # update the number of samples processed
-                self.logger.log('Number of samples', self.n_samples, self.n_iters, color='blue')
-                # Log parameters
-                for p in self.agent.parameters():
-                    self.logger.log(f"Parameter: {p.name}", p.data, self.n_iters, color='red')
-
-            # Update counters
-            self.n_epochs = info_sample['n_epochs']  # update the number of epochs completed
-            self.n_iters += 1
-        return
-
-    # Can be overridden by subclasses to implement specific sampling strategies
-    def sample(self, agents, verbose=False, **kwargs):
-        """ Sample a batch of data based on the proposed parameters. All proposals are evaluated on the same batch of inputs.
-
-        Args:
-            agents (list): A list of trace.Modules (proposed parameters) to evaluate.
-                **kwargs: Additional keyword arguments that may be used by the implementation.
-        """
-        samples = Samples(*self.train_sampler.sample(agents))  # create a Samples object to store the samples and the minibatch
-
-        # Log information about the sampling
-        scores = [ g.get_scores() for g in samples.samples]  # list of list of scores for each RolloutsGraph
-        scores = [item for sublist in scores for item in sublist]  # flatten the list of scores
-        log_info = {
-            'mean_score': np.mean(scores),
-            'n_epochs': self.train_sampler.n_epochs,
-        }
-        return samples, log_info
-
-    def log(self, info_log, prefix=""):
-        """ Log the information from the algorithm. """
-        for key, value in info_log.items():
-            try:
-                if value is not None:
-                    self.logger.log(f"{prefix}{key}", value, self.n_iters)
-            except Exception as e:
-                print(e)
-
-    def test(self, test_dataset, guide):
-        min_score = self.score_range[0]
-        # Test the agent's performance
-        test_score = self.evaluate(self.agent, guide, test_dataset['inputs'], test_dataset['infos'],
-                          min_score=min_score, num_threads=self.num_threads,
-                          description=f"Evaluating agent (iteration {self.n_iters})")  # and log
-        return {'test_score': test_score}
-
-    def save(self, save_path):
-        self.save_agent(save_path, self.n_iters)
-        # TODO save full state of self
-
-    # Unimplemented methods that should be implemented by subclasses
-    def update(self, samples=None, verbose=False, **kwargs):
-        """ Update the agent based on the provided samples.
-        Args:
-            samples (list): A list of samples from the previous iteration. If None, the agent's parameters are returned without updating.
-            verbose (bool, optional): Whether to print verbose output. Defaults to False.
-            **kwargs: Additional keyword arguments that may be used by the implementation.
-        Returns:
-            update_dict (dict of Parameter: Any): A dictionary containing the updated parameters of the agent.
-            proposals (list of trace.Module): A list of proposed parameters (trace.Module) after the update.
-            info_log (dict of str: Any): A dictionary containing logging information about the update process.
-
-        This method updates the agent's parameters based on samples of the training dataset and validation dataset (provided by self.get_validate_dataset).
-        In addition, it return new agents (proposals) that can be used for collecting data for the next iteration.
-        """
-        raise NotImplementedError("The update method should be implemented by subclasses.")
-        # return update_dict, proposals, info_log
+from opto.trainer.utils import async_run
+from opto.trainer.algorithms.basic_algorithms import batchify
+from opto.trainer.algorithms.priority_search.search_template import SearchTemplate, Samples
+from opto.trainer.algorithms.priority_search.utils import set_module_parameters, remap_update_dict, create_module_from_update_dict
 
 
 # TODO make this hashable?
